change datatype of id to int
* was changed datatype of Order#id from long to int

+++ src/main/java/me/kqlqk/shop/model/Order.java
@@ -21,7 +21,7 @@ public class Order {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id", insertable = false, updatable = false)
-    private long id; // TODO change to int
+    private int id;
 
     @ManyToOne
     @JoinColumn(name = "color_id", referencedColumnName = "id")
